PORTAL-6292 | @jkalberer | Add extraction for class members with the i18n object
