+++ tests/__fixtures__/testGetClassMember/customInstanceName.json
@@ -1,0 +1,10 @@
+{
+  "description": "test giving custom instance name to i18next",
+  "pluginOptions": {
+    "i18nextInstanceNames": ["pgm", "_"]
+  },
+  "expectValues": {
+    "key1": "some value",
+    "key2": "some value"
+  }
+}
